+++ flexmeasures/api/v3_0/tests/test_sensors_api.py
@@ -13,8 +13,105 @@ from flexmeasures.data.models.audit_log import AssetAuditLog
 from flexmeasures.data.schemas.sensors import SensorSchema
 from flexmeasures.data.models.generic_assets import GenericAsset
 from flexmeasures.tests.utils import QueryCounter
+from flexmeasures.utils.unit_utils import is_valid_unit
+
 
 sensor_schema = SensorSchema()
+
+
+@pytest.mark.parametrize(
+    "requesting_user, search_by, search_value, exp_sensor_name, exp_num_results, all_accessible, use_pagination",
+    [
+        (
+            "test_supplier_user_4@seita.nl",
+            "unit",
+            "°C",
+            "some temperature sensor",
+            2,
+            True,
+            False,
+        ),
+        (
+            "test_supplier_user_4@seita.nl",
+            "unit",
+            "m³/h",
+            "some gas sensor",
+            1,
+            True,
+            False,
+        ),
+        (
+            "test_supplier_user_4@seita.nl",
+            None,
+            None,
+            "some temperature sensor",
+            3,
+            True,
+            True,
+        ),
+        (
+            "test_supplier_user_4@seita.nl",
+            "filter",
+            "'some temperature sensor'",
+            "some temperature sensor",
+            1,
+            False,
+            False,
+        ),
+    ],
+    indirect=["requesting_user"],
+)
+def test_fetch_sensors(
+    client,
+    setup_api_test_data,
+    requesting_user,
+    search_by,
+    search_value,
+    exp_sensor_name,
+    exp_num_results,
+    all_accessible,
+    use_pagination,
+):
+    """
+    Retrieve all sensors.
+
+    Our user here is admin, so is allowed to see all sensors.
+    Pagination is tested only in passing, we should test filtering and page > 1
+    """
+    query = {search_by: search_value}
+
+    if use_pagination:
+        query["page"] = 1
+
+    if search_by == "unit":
+        query["unit"] = search_value
+    elif search_by == "filter":
+        query["filter"] = search_value
+
+    if all_accessible:
+        query["all_accessible"] = True
+
+    response = client.get(
+        url_for("SensorAPI:index"),
+        query_string=query,
+    )
+
+    print("Server responded with:\n%s" % response.json)
+    assert response.status_code == 200
+
+    if use_pagination:
+        assert isinstance(response.json["data"][0], dict)
+        assert is_valid_unit(response.json["data"][0]["unit"])
+        assert response.json["num-records"] == exp_num_results
+        assert response.json["filtered-records"] == exp_num_results
+    else:
+        assert isinstance(response.json, list)
+        assert is_valid_unit(response.json[0]["unit"])
+        assert response.json[0]["name"] == exp_sensor_name
+        assert len(response.json) == exp_num_results
+
+        if search_by == "unit":
+            assert response.json[0]["unit"] == search_value
 
 
 @pytest.mark.parametrize(